refactor(simulate): extract picker helpers to utils/simulationTestModeUtils
VOICE_ONLY_METRICS, NEVER_PICKABLE_TOPLEVEL, isTextCallDetail,
isHiddenPickerPath, translateDeepScenarioColumn move out of the .jsx file
so it exports only its component (Fast Refresh boundary). Test file moves
alongside them and renames from Helpers to Utils. Zero behaviour change.
